Moved SafeJSONEncoder into its own encoder file
